Deconflict php-fpm port from xdebug port.
Testing has identified this as the source of php-fpm processes becoming
hung even when a live debugging session is not active. A live session
should cause the connect back host to deviate from the default localhost
value and a non live session should cause no communication for xdebug.
However, experience has shown that something causes php-fpm processes
to become hung in a recv syscall which typically happens during bulk
receipt of many requests in a short time period which would normally
take a few seconds to fulfill. Deconflicting the ports resulted in the
inability to trigger the previous hanging behavior.